+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.2.3]
+
+- fix: deprecate `auth()`, use `setAuth()` instead [#107](https://github.com/supabase/postgrest-dart/pull/107)
+- fix: requests don't affect client's headers [#107](https://github.com/supabase/postgrest-dart/pull/107)
+
 ## [1.2.2]
 
 - fix: deprecate `returning` parameter of `.delete()` [#105](https://github.com/supabase/postgrest-dart/pull/105)